% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z2.R
\name{z2}
\alias{z2}
\title{Z2}
\usage{
z2(data)
}
\arguments{
\item{data}{The dataset of interest. Note that ROA should be in column number 3 and CAR in column number 4.}
}
\description{
Estimates rolling Z-scores with a fixed window length, k, set to 3. However, only the mean and standard deviation of ROA are computed on a rolling basis.
}
\keyword{data}
