% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z3.R
\name{z3}
\alias{z3}
\title{Z3}
\usage{
z3(data)
}
\arguments{
\item{data}{The dataset of interest. Note that ROA should be in column number 3 and CAR in column number 4.}
}
\description{
Estimates the Z-score by using ROA and CAR observations at every time t. If total years = T, you get T Z-scores in stead of T-k. Note that the standard deviation is estimated over the entire sample period.
}
\keyword{data}
