% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z4.R
\name{z4}
\alias{z4}
\title{Z4}
\usage{
z4(data)
}
\arguments{
\item{data}{The dataset of interest. Note that ROA should be in column number 3 and CAR in column number 4.}
}
\description{
Similar to Z3 in the sense that it is observed for every t. However, the standard deviation is computed differently. Here it is the absolute value of the deviation from the sample period mean of ROA for every t.
}
\keyword{data}
