% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z5.R
\name{z5}
\alias{z5}
\title{Z5}
\usage{
z5(data)
}
\arguments{
\item{data}{The dataset of interest. Note that ROA should be in column number 3 and CAR in column number 4.}
}
\description{
Estimates Z-scores with the overall sample period mean and sigma for ROA, whilst letting CAR be observed per time t. Result will be T Z-scores.
}
\keyword{data}
